itemstats: change mdmg type to float

+++ http-api/src/main/java/net/runelite/http/api/item/ItemEquipmentStats.java
@@ -51,7 +51,7 @@ public class ItemEquipmentStats
 
 	private int str;
 	private int rstr;
-	private int mdmg;
+	private float mdmg;
 	private int prayer;
 	private int aspeed;
 }
